refactor(extensions): move ExtensionPackage cluster into package.rs
Extract ExtensionPackage, CapabilityDescriptorSchemaMode, their impl
block, and the five private helpers used exclusively by them
(descriptors_match_except_schema, ensure_extension_root_matches,
extension_id_from_package_root, capability_descriptors_from_manifest,
invalid_package_root, descriptor_schema_ref) out of lib.rs into a new
sibling module. lib.rs drops from 815 to 546 lines.

Pure move: no logic, signature, or visibility changes. The external
path ironclaw_extensions::ExtensionPackage is preserved via pub use,
so no other crate is touched. Only now-dead imports were trimmed from
lib.rs to keep clippy at zero warnings.

Prophylactic — the file-size gate does not fire until 1500 lines, but
upcoming work adds a constructor to exactly this cluster and this is
the cheapest moment to split it.
